we need to unify tempfilenames here too.
